rename Images.tsx to ImagePairs
